+++ packages/widget/.storybook/preview.ts
@@ -1,3 +1,6 @@
 import "@navikt/ds-css";
 
+// Initialize mock consent module for Storybook
+import "./mocks/consentMock";
+
 export { previewAnnotations as default } from "../src/storybook/annotations";